+++ src/main.c
@@ -1,7 +1,7 @@
 #include <otter_slice.h>
 
 otter_slice_t
-otter_slice_init(void *ptr, size_t len) {
+otter_slice_init(void *ptr, unsigned long len) {
   otter_slice_t slice;
   slice.ptr = ptr;
   slice.len = len;
@@ -9,7 +9,7 @@ otter_slice_init(void *ptr, size_t len) {
 }
 
 otter_slice_result_t
-otter_slice_get(otter_slice_t slice, size_t index, void **out) {
+otter_slice_get(otter_slice_t slice, unsigned long index, void **out) {
   if (index >= slice.len) {
     return OTTER_SLICE_INDEX_OUT_OF_RANGE;
   }
@@ -20,7 +20,7 @@ otter_slice_get(otter_slice_t slice, size_t index, void **out) {
 }
 
 otter_slice_result_t
-otter_slice_slice(otter_slice_t slice, size_t start, size_t end, otter_slice_t *out) {
+otter_slice_slice(otter_slice_t slice, unsigned long start, unsigned long end, otter_slice_t *out) {
   otter_slice_t result;
 
   if (start >= slice.len || end > slice.len) {
